+++ Engine/src/Renderer/DX11/DX11_Buffer.h
@@ -8,15 +8,15 @@ namespace Alexio
 	class DX11_VertexBuffer : public VertexBuffer
 	{
 	public:
-		DX11_VertexBuffer(size_t size);
-		DX11_VertexBuffer(const void* data, size_t size);
+		DX11_VertexBuffer(uint32_t size);
+		DX11_VertexBuffer(const void* data, uint32_t size);
 
 		~DX11_VertexBuffer();
 
 		void Bind() const override;
 		void Unbind() const override;
 
-		void SetData(const void* data, size_t size) override;
+		void SetData(const void* data, uint32_t size) override;
 
 	private:
 		Microsoft::WRL::ComPtr<ID3D11Buffer> mBuffer;
@@ -25,7 +25,7 @@ namespace Alexio
 	class DX11_IndexBuffer : public IndexBuffer
 	{
 	public:
-		DX11_IndexBuffer(uint32_t* indices, size_t count);
+		DX11_IndexBuffer(uint32_t* indices, uint32_t count);
 		~DX11_IndexBuffer();
 
 		void Bind() override;
